+++ keycap_designer/__init__.py
@@ -5,3 +5,10 @@ except ImportError:
     __version_info__ = version = '0.0.dev1'
 
 __version__ = '.'.join(map(str, __version_info__))
+
+
+class InvalidDataError(Exception):
+    """
+    Something exists but it is invalid.
+    """
+    pass