+++ keycap_designer/jupyter.py
@@ -1,0 +1,2 @@
+from keycap_designer.manuscript import *  # noqa
+from keycap_designer._jupyter import show_inline, generate_pdf, change_default_dpi  # noqa